+++ bim2sim/meta_structure/bps.py
@@ -30,7 +30,7 @@ from bim2sim.decorators import cached_property
 from bim2sim.kernel import attribute, condition
 from bim2sim.meta_structure import ProductBased, RelationBased
 from bim2sim.kernel.units import ureg
-from bim2sim.task.common.inner_loop_remover import remove_inner_loops
+from bim2sim.tasks.common.inner_loop_remover import remove_inner_loops
 from bim2sim.utilities.common_functions import vector_angle, angle_equivalent
 from bim2sim.utilities.pyocc_tools import PyOCCTools
 from bim2sim.utilities.types import IFCDomain
