+++ bim2sim/plugins/PluginAixLib/bim2sim_aixlib/__init__.py
@@ -4,7 +4,7 @@ from ast import literal_eval
 from bim2sim.export.modelica import standardlibrary
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginAixLib.bim2sim_aixlib.models import AixLib
-from bim2sim.task import base, common, hvac
+from bim2sim.tasks import base, common, hvac
 from bim2sim.sim_settings import PlantSimSettings
