+++ bim2sim/plugins/PluginCFD/bim2sim_cfd/task/ifc2cfd.py
@@ -2,8 +2,8 @@ import os
 
 from bim2sim.decision import ListDecision, DecisionBunch, RealDecision
 from bim2sim.kernel.units import ureg
-from bim2sim.task.base import ITask
-from bim2sim.task.common import LoadIFC
+from bim2sim.tasks.base import ITask
+from bim2sim.tasks.common import LoadIFC
 
 
 class RunIFC2CFD(ITask):
@@ -14,7 +14,7 @@ class RunIFC2CFD(ITask):
 
     def run(self):
         if os.name != 'posix':
-            raise OSError("CFD task is only available for Linux systems")
+            raise OSError("CFD tasks is only available for Linux systems")
         self.logger.info("Running IFC2CFD")
 
         process_options = [
