+++ bim2sim/plugins/PluginEnergyPlus/bim2sim_energyplus/task/ep_idf_cfd_export.py
@@ -10,7 +10,7 @@ from OCC.Core.TopoDS import TopoDS_Shape
 from stl import mesh, stl
 
 from bim2sim.meta_structure.bps import SpaceBoundary
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances, \
     get_spaces_with_bounds
 from bim2sim.utilities.pyocc_tools import PyOCCTools
