+++ bim2sim/plugins/PluginEnergyPlus/bim2sim_energyplus/task/ep_run_simulation.py
@@ -1,7 +1,7 @@
 import subprocess
 
 from bim2sim.decision import BoolDecision, DecisionBunch
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 
 
 class RunEnergyPlusSimulation(ITask):
