+++ bim2sim/plugins/PluginLCA/bim2sim_lca/__init__.py
@@ -4,7 +4,7 @@ Holds logic to export LCA relevant information and quantities based on ifc data
 """
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginLCA.bim2sim_lca.task.export_lca import ExportLCA
-from bim2sim.task import common, bps
+from bim2sim.tasks import common, bps
 from bim2sim.sim_settings import LCAExportSettings
