+++ bim2sim/plugins/PluginTEASER/bim2sim_teaser/__init__.py
@@ -7,11 +7,11 @@ import inspect
 import bim2sim.plugins.PluginTEASER.bim2sim_teaser.task as teaser_task
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginTEASER.bim2sim_teaser.models import TEASER
-from bim2sim.task import common, bps, base
+from bim2sim.tasks import common, bps, base
 from bim2sim.sim_settings import BuildingSimSettings, Setting
 from bim2sim.utilities.types import LOD
 
-from bim2sim.task.bps.bind_tz import CombineThermalZones
+from bim2sim.tasks.bps.bind_tz import CombineThermalZones
 
 
 class TEASERSimSettings(BuildingSimSettings):
