+++ bim2sim/plugins/PluginTEASER/bim2sim_teaser/task/export_teaser.py
@@ -13,7 +13,7 @@ from teaser.logic.buildingobjects.buildingphysics.window import Window
 from teaser.project import Project
 
 from bim2sim.meta_structure.__init__ import ProductBased
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
