+++ bim2sim/project.py
@@ -13,7 +13,7 @@ import configparser
 
 from bim2sim.decision import ListDecision, DecisionBunch, save, load
 from bim2sim import log
-from bim2sim.task.base import Playground
+from bim2sim.tasks.base import Playground
 from bim2sim.plugins import Plugin, load_plugin
 from bim2sim.utilities.common_functions import all_subclasses
 from bim2sim.sim_settings import BaseSimSettings, AutoSettingNameMeta
