+++ bim2sim/tasks/bps/bind_tz.py
@@ -2,7 +2,7 @@ import inspect
 
 from bim2sim.decision import ListDecision, DecisionBunch
 from bim2sim.kernel.aggregation import AggregatedThermalZone
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
 from bim2sim.utilities.types import LOD
 
@@ -10,9 +10,9 @@ from bim2sim.utilities.types import LOD
 class CombineThermalZones(ITask):
     """Combine thermal zones to reduce the amount of thermal zones.
 
-    As the zoning of simulation models is a time-consuming task we decided to
-    automate it with the task.
-    This task will combine multiple thermal zones into one zone based on the
+    As the zoning of simulation models is a time-consuming tasks we decided to
+    automate it with the tasks.
+    This tasks will combine multiple thermal zones into one zone based on the
     criteria selected in the simulation type settings and the decisions made.
     We do this by giving the user multiple criteria to select from:
         * External/Internal
