+++ bim2sim/tasks/bps/disaggr_creation.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from bim2sim.decorators import cached_property
 from bim2sim.kernel import attribute
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
 from bim2sim.utilities.types import LOD
