+++ bim2sim/tasks/bps/enrich_use_cond.py
@@ -2,7 +2,7 @@ from typing import Union, Dict
 
 from bim2sim.decision import ListDecision, DecisionBunch
 from bim2sim.meta_structure.bps import ThermalZone
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import get_usage_dict, get_pattern_usage
