+++ bim2sim/tasks/bps/prepare.py
@@ -1,6 +1,6 @@
 from bim2sim.decision import BoolDecision, DecisionBunch
 from bim2sim.meta_structure.bps import Slab, GroundFloor, Floor, Roof
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
