+++ bim2sim/tasks/bps/sb_creation.py
@@ -13,7 +13,7 @@ from bim2sim.meta_structure.bps import SpaceBoundary, ExtSpatialSpaceBoundary, \
     ThermalZone, Window, Door
 from bim2sim.kernel.finder import TemplateFinder
 from bim2sim.kernel.units import ureg
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.sim_settings import BaseSimSettings
 
 logger = logging.getLogger(__name__)
