+++ bim2sim/tasks/bps/visualize_spaces.py
