+++ bim2sim/tasks/common/bind_storeys.py
@@ -1,4 +1,4 @@
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
