+++ bim2sim/tasks/common/weather.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from bim2sim.decision import ListDecision, DecisionBunch
-from bim2sim.task.base import ITask
+from bim2sim.tasks.base import ITask
 
 
 class Weather(ITask):
