+++ test/unit/kernel/task/test_dead_ends.py
@@ -2,7 +2,7 @@ import unittest
 
 from bim2sim.meta_structure import hvac
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
-from bim2sim.task.hvac import dead_ends
+from bim2sim.tasks.hvac import dead_ends
 from test.unit.kernel.helper import SetupHelperHVAC
